requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-functions==1.17.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.2
diskcache==5.6.3
Django==4.2.7
dnspython==2.4.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
image==1.5.33
jsonpatch==1.33
jsonpointer==2.4
jsons==1.6.3
langchain==0.0.336
langsmith==0.0.64
llama_cpp_python==0.2.20
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.27.8
packaging==23.2
Pillow==10.1.0
pinecone-client==2.2.4
pydantic==2.5.1
pydantic_core==2.14.3
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pytesseract==0.3.10
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
typish==1.9.3
tzdata==2023.3
urllib3==2.0.7
win32-setctime==1.1.0
yarl==1.9.2
